Mental-Wellness-database/requirements.txt
langchain
flask
pymongo
python-dotenv
pinecone
llm
langchain-google-genai
Flask-Cors
pydantic[email]

